feat(backend,nextjs,utils): Remove interstitial
