+++ packages/backend/src/tokens/errors.ts
@@ -22,8 +22,6 @@ export enum TokenVerificationErrorReason {
   RemoteJWKMissing = 'jwk-remote-missing',
 
   JWKFailedToResolve = 'jwk-failed-to-resolve',
-
-  RemoteInterstitialFailedToLoad = 'interstitial-remote-failed-to-load',
 }
 
 export enum TokenVerificationErrorAction {
