+++ packages/backend/src/tokens/factory.ts
@@ -1,7 +1,4 @@
-import type { ApiClient } from '../api';
 import { mergePreDefinedOptions } from '../util/mergePreDefinedOptions';
-import type { LoadInterstitialOptions } from './interstitial';
-import { loadInterstitialFromLocal } from './interstitial';
 import type { AuthenticateRequestOptions } from './request';
 import { authenticateRequest as authenticateRequestOriginal, debugRequestState } from './request';
 
@@ -36,11 +33,9 @@ const defaultOptions = {
 
 export type CreateAuthenticateRequestOptions = {
   options: BuildTimeOptions;
-  apiClient: ApiClient;
 };
 
 export function createAuthenticateRequest(params: CreateAuthenticateRequestOptions) {
-  const { apiClient } = params;
   const buildTimeOptions = mergePreDefinedOptions(defaultOptions, params.options);
 
   const authenticateRequest = (options: RunTimeOptions) => {
@@ -56,18 +51,8 @@ export function createAuthenticateRequest(params: CreateAuthenticateRequestOptio
     });
   };
 
-  const localInterstitial = (options: Omit<LoadInterstitialOptions, 'apiUrl'>) => {
-    const runTimeOptions = mergePreDefinedOptions(buildTimeOptions, options);
-    return loadInterstitialFromLocal({ ...options, ...runTimeOptions });
-  };
-
-  // TODO: Replace this function with remotePublicInterstitial
-  const remotePrivateInterstitial = () => apiClient.interstitial.getInterstitial();
-
   return {
     authenticateRequest,
-    localInterstitial,
-    remotePrivateInterstitial,
     debugRequestState,
   };
 }
