+++ packages/backend/src/tokens/index.ts
@@ -3,6 +3,5 @@ export { AuthStatus } from './authStatus';
 export type { RequestState } from './authStatus';
 export * from './errors';
 export * from './factory';
-export { loadInterstitialFromLocal } from './interstitial';
 export { debugRequestState } from './request';
 export type { AuthenticateRequestOptions, OptionalVerifyTokenOptions } from './request';
