+++ packages/backend/src/util/path.ts
@@ -1,11 +1,2 @@
-const SEPARATOR = '/';
-const MULTIPLE_SEPARATOR_REGEX = new RegExp(SEPARATOR + '{1,}', 'g');
-
-type PathString = string | null | undefined;
-
-export function joinPaths(...args: PathString[]): string {
-  return args
-    .filter(p => p)
-    .join(SEPARATOR)
-    .replace(MULTIPLE_SEPARATOR_REGEX, SEPARATOR);
-}
+// TODO: @nikos refactor this to use the shared package directly
+export { joinPaths } from '@clerk/shared/url';
